The pipeTableExists function was previously used, however in some versions of mysql case sensitive problems were reported in the function names. Then, we replace it with a query.
Version 5.7. * Of mysql sets the sql_mode parameter to STRICT_TRANS_TABLES by default. This makes it impossible to use a datetime value of '000-00-00 00:00:00'

To solve the problem, just remove the STRICT_TRANS_TABLES from the sql_mode parameter. However, some users do not have access to mysql.ini, or permission to set in real time, so we created a workaround, changing the example record in the bbd_tbmessage table.